Move /election/<election>/constituencies/unlocked to /elections/<election>/unlocked/
