+++ ynr/apps/candidates/views/constituencies.py
@@ -989,47 +989,6 @@ class ConstituencyLockView(ElectionMixin, GroupRequiredMixin, View):
             raise ValidationError(message)
 
 
-class ConstituenciesUnlockedListView(ElectionMixin, TemplateView):
-    template_name = "candidates/constituencies-unlocked.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        total_constituencies = 0
-        total_locked = 0
-        keys = ("locked", "unlocked")
-        for k in keys:
-            context[k] = []
-        postextraelections = (
-            PostExtraElection.objects.filter(election=self.election_data)
-            .select_related("post")
-            .all()
-        )
-        for postextraelection in postextraelections:
-            total_constituencies += 1
-            if postextraelection.candidates_locked:
-                context_field = "locked"
-                total_locked += 1
-            else:
-                context_field = "unlocked"
-            context[context_field].append(
-                {
-                    "id": postextraelection.post.slug,
-                    "name": postextraelection.post.short_label,
-                }
-            )
-        for k in keys:
-            context[k].sort(key=lambda c: c["name"])
-        context["total_constituencies"] = total_constituencies
-        context["total_left"] = total_constituencies - total_locked
-        if total_constituencies > 0:
-            context["percent_done"] = (
-                100 * total_locked
-            ) // total_constituencies
-        else:
-            context["percent_done"] = 0
-        return context
-
-
 class ConstituencyRecordWinnerView(ElectionMixin, GroupRequiredMixin, FormView):
 
     form_class = ConstituencyRecordWinnerForm
